+++ dev/tests/static/testsuite/Magento/Test/Legacy/_files/obsolete_config_nodes.php
@@ -31,4 +31,5 @@ return array(
     '/config/global/cache/types/*/tags' => 'use /config/global/cache/types/*/class node instead',
     '/config/global/disable_local_modules' => '',
     '/config/*/layout' => 'Use convention for layout files placement instead of configuration',
+    '/config/vde' => 'Was moved to di',
 );