+++ dev/tests/static/testsuite/Magento/Test/Legacy/_files/obsolete_constants.php
@@ -133,4 +133,6 @@ return array(
     array('XML_STORE_ROUTERS_PATH', 'Mage_Core_Controller_Varien_Front'),
     array('XML_PATH_THEME'),
     array('XML_PATH_ALLOW_DUPLICATION'),
+    array('XML_PATH_FRONT_NAME'),
+    array('XML_PATH_DISABLED_CACHE_TYPES'),
 );